+++ utoipa-gen/CHANGELOG.md
@@ -37,6 +37,7 @@
 
 ### Changed
 
+* Fix broken tests
 * Fix typos
 * Update `utoipa-config` version
 * Remove commit commit id from changelogs (https://github.com/juhaku/utoipa/pull/1077)